plot/plotRutils.c: Multiplication result converted to larger type
Fix code scanning alert no. 54: Multiplication result converted to larger type (#56)

+++ plot/plotRutils.c
@@ -519,7 +519,7 @@ PlotDumpRaster(raster, file)
     int count;
 
     count = write(fileno(file), (char *) raster->ras_bits,
-	    raster->ras_bytesPerLine*raster->ras_height);
+	    (size_t)raster->ras_bytesPerLine * raster->ras_height);
     if (count < 0)
     {
 	TxError("I/O error in writing raster file:  %s.\n",
